[BE] FEAT: 해몽 생성 API 적용 및 이미지 생성 API 와 경로 통합 #30
