+++ backend/src/dreamdiary/dreamdiary.controller.ts
@@ -4,7 +4,9 @@ import {
   ForbiddenException,
   InternalServerErrorException,
   Logger,
+  Param,
   Post,
+  Put,
   UseGuards,
   ValidationPipe,
 } from '@nestjs/common';
@@ -66,5 +68,16 @@ export class DreamDiaryController {
       }
       throw new InternalServerErrorException(err.message);
     }
+  }
+
+  @ApiOperation({
+    summary: '해몽 생성',
+    description: '해당 다이어리 내용의 해몽을 저장합니다.',
+  })
+  @Put(':id/interpretation')
+  async createDreamInterpretation(@Param('id') diaryId: number) {
+    const interpretation =
+      await this.dreamDiaryService.createDreamInterpretation(diaryId);
+    return { interpretation };
   }
 }
