+++ backend/src/dreamdiary/dreamdiary.service.ts
@@ -1,17 +1,22 @@
 import { ForbiddenException, Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { InjectRepository } from '@nestjs/typeorm';
 import { Configuration, OpenAIApi } from 'openai';
 import { GeneratedImagesResponseDto } from 'src/dto/generated.images.response.dto';
+import { DreamDiary } from 'src/entities/dream.diary.entity';
 import { UserService } from 'src/user/user.service';
 import { OpenAIService } from 'src/util/openai.service';
+import { Repository } from 'typeorm';
 
 @Injectable()
 export class DreamDiaryService {
   private readonly logger = new Logger(DreamDiaryService.name);
 
   constructor(
-    private readonly oepnAIService: OpenAIService,
+    private readonly openAIService: OpenAIService,
     private readonly userService: UserService,
+    @InjectRepository(DreamDiary)
+    private readonly dreamDiaryRepository: Repository<DreamDiary>,
   ) {}
 
   async createDreamDiaryImages(
@@ -66,5 +71,23 @@ export class DreamDiaryService {
       maxFreeGenerateCount: maxRequestCount,
       generatedImages: images,
     } as GeneratedImagesResponseDto;
+  }
+
+  async createDreamInterpretation(diaryId: number) {
+    this.logger.debug(`Called ${this.createDreamInterpretation.name}`);
+
+    const dreamDiary = await this.dreamDiaryRepository.findOne({
+      where: { diaryId: diaryId },
+    });
+
+    const interpretation = await this.openAIService.createDreamInterpretation(
+      dreamDiary.title,
+      dreamDiary.content,
+    );
+
+    dreamDiary.interpretation = interpretation;
+
+    await this.dreamDiaryRepository.save(dreamDiary);
+    return interpretation;
   }
 }
