+++ backend/src/util/openai.service.ts
@@ -10,8 +10,6 @@ export class OpenAIService {
     const config = new Configuration({
       apiKey: this.configService.get<string>('openai.apiKey'),
       organization: this.configService.get<string>('openai.organizationID'),
-      // apiKey: process.env.OPENAI_API_KEY,
-      // organization: process.env.OPENAI_ORGANIZATION_ID,
     });
     this.openAIApi = new OpenAIApi(config);
   }
@@ -42,19 +40,26 @@ export class OpenAIService {
   /**
    * OpenAI의 Davinci 엔진을 이용해 해몽을 생성합니다.
    *
-   * @param prompt 해몽을 위한 프롬프트
+   * @param title 꿈일기 제목
+   * @param content 꿈일기 내용
    * @returns 생성된 해몽 결과
    */
-  async createDreamInterpretation(content: string): Promise<string> {
+  async createDreamInterpretation(
+    title: string,
+    content: string,
+  ): Promise<string> {
     this.logger.debug(`Called ${this.createDreamInterpretation.name}`);
 
     const response = await this.openAIApi.createCompletion({
       model: 'text-davinci-003',
       prompt:
+        '제목은 ' +
+        title +
+        '이고 내용은 ' +
         content +
-        '\n\n이 꿈에 대한 세부적인 설명은 필요 없이 해몽을 알려줘. 대답은 할 필요 없이 "해몽 : "으로 시작해.',
+        '\n\n이것은 꿈일기이고 이 꿈일기의 꿈에 대한 해몽을 알려줘. 꿈에 대한 요약과 세부적인 설명은 필요 없어. 대답은 할 필요 없이 "해몽 : "으로 시작해.',
       max_tokens: 1000,
-      temperature: 0.4,
+      temperature: 0.3,
       n: 1,
     });
